fix cyclical dependencies
